kichobot6

+++ bot/middlewares/bot/callbackAntiSpamMiddleware.js
@@ -1,5 +1,5 @@
 import i18next from "i18next";
-import {bot, userLastRequest} from "../../app.js";
+import {bot, userLastRequest, userWarningSent} from "../../app.js";
 import log from "../../logging/logging.js";
 import userService from "../../services/userService.js";
 
@@ -13,20 +13,23 @@ export async function callbackAntiSpamMiddleware(call, next) {
             const timeDiff = currentTime - userLastRequest[userId];
 
             if (timeDiff < 750) {
-                // Отправляем уведомление о спаме пользователю
-                const user_language = await userService.getUserLanguage(call.message.chat.id)
-                const msg_text = i18next.t('antispam', {lng:user_language})
+                if (!userWarningSent[userId] || (currentTime - userWarningSent[userId] > 5000)) {
+                    userWarningSent[userId] = currentTime;
+                    // Отправляем уведомление о спаме пользователю
+                    const user_language = await userService.getUserLanguage(call.message.chat.id)
+                    const msg_text = i18next.t('antispam', {lng:user_language})
 
-                await bot.answerCallbackQuery(call.id, {text: msg_text, show_alert: false})
-                    .catch(async (e) => {
-                        try {
-                            log.error(`User ${call.message.chat.id} got an error в коллбек антиспам мидлваре` + e.message, {stack: e.stack})
-                            await bot.deleteMessage(call.message.chat.id, call.message.message_id)
-                            await bot.sendMessage(call.message.chat.id, "⚠️Произошла ошибка! Попробуйте получить ваше меню снова.")
-                        } catch (e) {
-                            log.error(`User ${call.message.chat.id} got an double!!! error в коллбек антиспам мидлваре` + e.message, {stack: e.stack})
-                        }
-                    })
+                    await bot.answerCallbackQuery(call.id, {text: msg_text, show_alert: false})
+                        .catch(async (e) => {
+                            try {
+                                log.error(`User ${call.message.chat.id} got an error в коллбек антиспам мидлваре` + e.message, {stack: e.stack})
+                                await bot.deleteMessage(call.message.chat.id, call.message.message_id)
+                                await bot.sendMessage(call.message.chat.id, "⚠️Произошла ошибка! Попробуйте получить ваше меню снова.")
+                            } catch (e) {
+                                log.error(`User ${call.message.chat.id} got an double!!! error в коллбек антиспам мидлваре` + e.message, {stack: e.stack})
+                            }
+                        })
+                }
                 // Пропускаем выполнение хендлера
                 return;
             }
